==> code/03_extract_prices_yahoo.ipynb <==
## 3 - Extract Prices
> Get price change for all stocks on document's release date

### 3.0 - Import Libraries

import pandas as pd
import numpy as np
import time
import random
import unicodedata
import re
import datetime
import os
from nltk.tokenize import RegexpTokenizer
from sklearn.preprocessing import MultiLabelBinarizer
import ast
import requests
from bs4 import BeautifulSoup
import pandas_datareader.data as web
import pandas_market_calendars as mcal
from tqdm import tqdm

# Set max columns and width with pandas
pd.set_option('display.max_colwidth', -1)
pd.set_option('display.max_columns', 500)

# Clear future and deprecation warnings
import warnings
warnings.filterwarnings("ignore", category=DeprecationWarning)
warnings.filterwarnings("ignore", category=FutureWarning)

### 3.1 Read in Data

df = pd.read_csv('../clean_data/concat.csv')
df.drop(columns=['index'], axis=1, inplace=True)

### 3.2 Drop Duplicates

# Check shape before
df.shape

# Override DataFrame dropping duplicate values
df = df.drop_duplicates(subset=['corpus', 'release_date'])

# Check shape after
df.shape

# Check DataFrame's first row
df.head(1)

#### Drop Columns `doc_link`, `complete_file_link`, `file_description`, `filing_doc`, `date`

df.drop(columns=['doc_link', 'complete_file_link', 'file_description', 'filing_doc', 'date'], inplace=True)
df.head(1)

#### Check for Null Values and Drop Them

df.isnull().sum()

df.dropna(inplace=True)

### 3.3 Create `tokens` column - Tokenize `corpus`

# Create function to tokenize 'body'
def tokenizer(x):
    tokenizer = RegexpTokenizer(r'\w+')
    return tokenizer.tokenize(x)

df['tokens'] = df['corpus'].progress_map(tokenizer)

### 3.4 Create `length_corpus` column

# Create column with corpus lenght
df['length_corpus'] = df['corpus'].progress_map(lambda x: len(x))

### 3.5 Create `items` Columns
>Learn More: [8-K Filings Item Types](www.sec.gov/fast-answers/answersform8khtm.html)

# Define function to extract document types
def extract_item(text):
    '''Extracts Types of Documents in each Filing'''
    # Compile Regular Expression
    exp = re.compile("Item+ +\d+[\:,\.]+\d+\d")
    
    # Create list with all items found
    item_list = re.findall(exp, text)
    item_list = [i.replace(' ', '').strip() for i in item_list]
    
    # Return list
    return item_list

# Create column with items types in each document
df['items'] = df['corpus'].progress_map(extract_item)

#### Convert `release_date` to Datetime

df['release_date'] = pd.to_datetime(df['release_date'])

### 3.6 - Get Stock Prices from Yahoo Finance
> _Yahoo! Finance is a media property that is part of Yahoo!'s network. It provides financial news, data and commentary including stock quotes, press releases, financial reports, and original content. It also offers some online tools for personal finance management._ [Learn More](https://finance.yahoo.com/)

> _Remote Data Access with Pandas DataReader_ [Documentation](https://pandas-datareader.readthedocs.io/en/latest/)

#### Define Function to Get Prices

def get_prices(ticker_list, start_date, end_date):
    '''
    Get Stock Prices from Yahoo Finance for Specific Time Frame
    Inputs = List of Tickers, Start Date and End Date
    Output = DataFrame
    '''
    # Instantiate DataFrame to house prices
    prices_df = pd.DataFrame()

    # Loop through all companies in the data
    for ticker in tqdm(ticker_list):

        # Get prices
        ticker_df = web.DataReader(ticker ,'yahoo', start_date, end_date)

        # Create columns with company's ticker
        ticker_df['company'] = ticker

        # Concatenate DataFrame with prices_df
        prices_df = pd.concat([prices_df, ticker_df], axis=0)
    
    # Remove data from index
    prices_df.reset_index(inplace=True)
    
    # Lower columns names
    prices_df.columns = [column.lower() for column in prices_df.columns]

    # Return DataFrame with all companies prices for requested timeframe
    return prices_df

prices_df = get_prices(df['company'].unique(), '2010-01-01', '2019-12-31')

prices_df.head(1)

#### Define Functions to Calculate Price Change

# Get list of holidays for the New York Stock Exchange
holidays = mcal.get_calendar('NYSE').holidays().holidays

def date_check(release_date):
    '''Checks for Holidays and Weekends'''
    while release_date.isoweekday() > 5 or release_date.date() in holidays:
        release_date = release_date + datetime.timedelta(days=-1)
    return release_date

def get_percentage_change(ticker, release_date, dataframe):    
    '''Calculates Price Change for a Security Given Document Release Date'''
    # define market hours
    market_open = release_date.replace(hour=9,minute=30,second=0)
    market_close = release_date.replace(hour=16,minute=0,second=0)
    
    try:
        # release_date BEFORE market open
        if release_date < market_open:

            # Get open price
            open_price = dataframe.loc[(dataframe['company'] == ticker) & \
                          (dataframe['date'] == release_date \
                                                .strftime('%m/%d/%Y').replace('/','-'))]['open'].item()

            # Get close price
            close_price = dataframe.loc[(dataframe['company'] == ticker) & \
                          (dataframe['date'] == date_check(release_date + datetime.timedelta(days=-1)) \
                                                .strftime('%m/%d/%Y').replace('/','-'))]['close'].item()

            percentage_change = (open_price - close_price) / close_price

        # release_date AFTER market close
        elif release_date > market_close:
            
            # Get open price
            open_price = dataframe.loc[(dataframe['company'] == ticker) & \
                          (dataframe['date'] == date_check(release_date + datetime.timedelta(days=+1))\
                                                .strftime('%m/%d/%Y').replace('/','-'))]['open'].item()

            # Get close price
            close_price = dataframe.loc[(dataframe['company'] == ticker) & \
                          (dataframe['date'] == release_date \
                                                .strftime('%m/%d/%Y').replace('/','-'))]['close'].item()

            # Percentage change if release_date > market_close
            percentage_change = (open_price - close_price) / close_price

        # release_date during trading hours
        else:
            # Get open price
            open_price = dataframe.loc[(dataframe['company'] == ticker) & \
                          (dataframe['date'] == release_date \
                                                .strftime('%m/%d/%Y').replace('/','-'))]['open'].item()

            # Get close price
            close_price = dataframe.loc[(dataframe['company'] == ticker) & \
                          (dataframe['date'] == release_date \
                                                .strftime('%m/%d/%Y').replace('/','-'))]['close'].item()

            percentage_change = (close_price - open_price) / open_price
            
    except:
        percentage_change = np.nan
        
    return percentage_change

#### Create Column with Percentage Change ( Stocks )
> _Explain Step_

df['stock_change'] = [get_percentage_change(df['company'][i],
                                            df['release_date'][i],
                                            prices_df) for i in tqdm(df.index)]

#### Create Column with Percentage Change for The S&P 500 Index ( GSPC )
> _The S&P 500, or just the S&P, is a stock market index that measures the stock performance of 500 large companies listed on stock exchanges in the United States. It is one of the most commonly followed equity indices, and many consider it to be one of the best representations of the U.S. stock market.The average annual total return of the index, including dividends, since inception in 1926 has been 9.8%; however, there were several years where the index declined over 30%.The index has posted annual increases 70% of the time._ [Learn More](https://en.wikipedia.org/wiki/S%26P_500_Index)

sp500_vix_df = get_prices(['^GSPC', '^VIX'], '2010-01-01', '2019-12-31')

sp500_vix_df.head(1)

df['index_change'] = [get_percentage_change('^GSPC',
                                        df['release_date'][i],
                                        sp500_vix_df) for i in tqdm(df.index)]

#### Create Column with Normalized Percentage Change
> _Explain Step_

df['normalized_change'] = df['stock_change'] - df['index_change']

#### Create Column with Percentage Change for The CBOE Volatility Index ( VIX )
> _The CBOE Volatility Index, known by its ticker symbol VIX, is a popular measure of the stock market's expectation of volatility implied by S&P 500 index options. It is calculated and disseminated on a real-time basis by the Chicago Board Options Exchange (CBOE), and is commonly referred to as the fear index or the fear gauge._ [Learn More](https://en.wikipedia.org/wiki/VIX)

def get_vix(release_date, dataframe):
    
    # define market hours
    market_open = release_date.replace(hour=9,minute=30,second=0)
    market_close = release_date.replace(hour=16,minute=0,second=0)
    
    try:
        if release_date < market_open:
            vix_price = dataframe.loc[(dataframe['company'] == '^VIX') & \
                                  (dataframe['date'] == date_check(release_date + datetime.timedelta(days=-1)) \
                                                        .strftime('%m/%d/%Y').replace('/','-'))]['close'].item()


    
        elif release_date > market_open:
            vix_price = dataframe.loc[(dataframe['company'] == '^VIX') & \
                                  (dataframe['date'] == release_date \
                                                        .strftime('%m/%d/%Y').replace('/','-'))]['close'].item()
        
        else:
            vix_price = dataframe.loc[(dataframe['company'] == '^VIX') & \
                                  (dataframe['date'] == release_date \
                                                        .strftime('%m/%d/%Y').replace('/','-'))]['open'].item()
    
    except:
        vix_price = np.nan
        
    return vix_price

df['vix'] = [get_vix(df['release_date'][i], sp500_vix_df) for i in tqdm(df.index)]

df.head(1)

#### Get Historical Movements

def get_historical(ticker, release_date, period, dataframe=prices_df):
        
    try:
        # Percentage change for the last week        
        if period == 'w':
            start = dataframe.loc[(dataframe['company'] == ticker) & \
                                  (dataframe['date'] == date_check(release_date + datetime.timedelta(weeks=-1)) \
                                                        .strftime('%m/%d/%Y').replace('/','-'))]['close'].item()
            end = dataframe.loc[(dataframe['company'] == ticker) & \
                          (dataframe['date'] == date_check(release_date + datetime.timedelta(days=-1)) \
                                                .strftime('%m/%d/%Y').replace('/','-'))]['close'].item()

        # Percentage change for the last month
        elif period == 'm':
            start = dataframe.loc[(dataframe['company'] == ticker) & \
                                  (dataframe['date'] == date_check(release_date + datetime.timedelta(weeks=-4)) \
                                                        .strftime('%m/%d/%Y').replace('/','-'))]['close'].item()
            end = dataframe.loc[(dataframe['company'] == ticker) & \
                                (dataframe['date'] == date_check(release_date + datetime.timedelta(days=-1)) \
                                                      .strftime('%m/%d/%Y').replace('/','-'))]['close'].item()

        # Percentage change for the last quarter
        elif period == 'q':
            start = dataframe.loc[(dataframe['company'] == ticker) & \
                                  (dataframe['date'] == date_check(release_date + datetime.timedelta(weeks=-12)) \
                                                        .strftime('%m/%d/%Y').replace('/','-'))]['close'].item()
            end = dataframe.loc[(dataframe['company'] == ticker) & \
                                (dataframe['date'] == date_check(release_date + datetime.timedelta(days=-1)) \
                                                      .strftime('%m/%d/%Y').replace('/','-'))]['close'].item()

        # Percentage change for the last year
        elif period == 'y':
            start = dataframe.loc[(dataframe['company'] == ticker) & \
                                  (dataframe['date'] == date_check(release_date + datetime.timedelta(weeks=-52)) \
                                                        .strftime('%m/%d/%Y').replace('/','-'))]['close'].item()
            end = dataframe.loc[(dataframe['company'] == ticker) & \
                                (dataframe['date'] == date_check(release_date + datetime.timedelta(days=-1)) \
                                                      .strftime('%m/%d/%Y').replace('/','-'))]['close'].item()
    except:
        return np.nan
            
    return (end - start) / start

#### Week - 1

df['hist_week'] = [get_historical(df['company'][i],
                                  df['release_date'][i],
                                  period='w',
                                  dataframe=prices_df) for i in tqdm(df.index)]

#### Month - 1

df['hist_month'] = [get_historical(df['company'][i],
                                  df['release_date'][i],
                                  period='m',
                                  dataframe=prices_df) for i in tqdm(df.index)]

#### Quarter - 1

df['hist_quarter'] = [get_historical(df['company'][i],
                                  df['release_date'][i],
                                  period='q',
                                  dataframe=prices_df) for i in tqdm(df.index)]

#### Year - 1

df['hist_year'] = [get_historical(df['company'][i],
                                  df['release_date'][i],
                                  period='y',
                                  dataframe=prices_df) for i in tqdm(df.index)]

df.head(1)

### 3.7 - Drop Rows with Errors

df.shape

df.isnull().sum()

df.dropna(axis=0, inplace=True)
df.reset_index(drop=True, inplace=True)

df.shape

### 3.8 Create `movement` Column
> Based on the normalized change column classify observations as being greater than 1% ( Up ), less than -1% ( Down ) or in between -1% and 1% ( Neutral ).

df['movement'] = df['normalized_change'].map(lambda x: 'down' if x < -.01 else ('up' if x > .01 else 'neutral'))

df.head(1)

#### Save DataFrame

df.to_csv('../clean_data/corpus.csv', index_label=False)
